+++ packages/demo/__stories__/index.tsx
@@ -1,5 +1,5 @@
 import { HtmlRenderer } from '@splish-me/editor-html-renderer'
-import { Button, ButtonGroup } from '@splish-me/editor-ui-plugin-sidebar'
+import { Button, ButtonGroup } from '@splish-me/editor-ui'
 import { storiesOf } from '@storybook/react'
 import * as R from 'ramda'
 import * as React from 'react'