+++ packages/demo/__stories__/sidebar.tsx
@@ -1,4 +1,4 @@
-import { Sidebar } from '@splish-me/editor-ui-sidebar'
+import { Sidebar } from '@splish-me/editor-ui'
 import { storiesOf } from '@storybook/react'
 import * as React from 'react'