+++ packages/demo/src/mock-broken-content.plugin.tsx
@@ -1,5 +1,5 @@
 import { PluginEditorProps } from '@splish-me/editor'
-import { renderIntoSidebar, Text } from '@splish-me/editor-ui-plugin-sidebar'
+import { renderIntoSidebar, Text } from '@splish-me/editor-ui'
 import * as React from 'react'
 
 export class MockBrokenContent extends React.Component<