+++ packages/demo/src/mock-wrapper.plugin.tsx
@@ -5,7 +5,7 @@ import {
   Plugin,
   PluginEditorProps
 } from '@splish-me/editor'
-import { renderIntoSidebar, Text } from '@splish-me/editor-ui-plugin-sidebar'
+import { renderIntoSidebar, Text } from '@splish-me/editor-ui'
 import * as React from 'react'
 
 export class MockWrapper extends React.Component<