+++ packages/plugin-text-plugin-link/src/index.tsx
@@ -3,7 +3,7 @@ import {
   NodeRendererProps,
   TextPlugin
 } from '@splish-me/editor-plugin-text-plugin'
-import { renderIntoSidebar, Input } from '@splish-me/editor-ui-plugin-sidebar'
+import { renderIntoSidebar, Input } from '@splish-me/editor-ui'
 import { debounce } from 'lodash'
 import { Change, Inline, Data, InlineJSON } from 'slate'
 import * as React from 'react'