+++ scripts/changelog.ts
@@ -20,6 +20,7 @@ async function exec(): Promise<void> {
         '@splish-me/editor replaces @splish-me/editor-core, @splish-me/editor-core-contexts, @splish-me/editor-core-document and @splish-me/editor-core-types',
         '@splish-me/editor-html-renderer replaces @splish-me/editor-core-html-renderer',
         '@splish-me/editor-renderer replaces @splish-me/editor-core-renderer',
+        '@splish-me/editor-ui replaces @splish-me/editor-ui-add-sidebar, @splish-me/editor-ui-mode-toolbar, @splish-me/editor-ui-plugin-sidebar and @splish-me/editor-ui-sidebar',
         "Don't provide `module` entry anymore"
       ],
       added: ['Pass plugins as dictionary'],